feat: implement LSP features for pyproject and Tombi extensions
- Added support for various LSP features in the pyproject extension, including code actions, completion, document links, hover, inlay hints, and navigation.
- Introduced enums and feature trees for managing feature toggles and defaults.
- Refactored the Tombi extension to utilize the new LSP feature structure, consolidating related functionality and improving organization.
- Each feature now includes methods to check if they are enabled and to access specific feature options.
